Former-commit-id: 48e954e8e43787c3bbebd4468e76cb8c9d37df7d
Former-commit-id: 4e361ea5bf64ef1ce34251c70569185e996326f2
Former-commit-id: 08cadb542536818467e894d690aeac87c90aa9f5

+++ airesources/Java/InitPackage.java
@@ -1,5 +1,5 @@
 
 public class InitPackage{
-  public short playerTag;
+  public int playerTag;
   public Map map;
 }
