+++ airesources/Java/Location.java
@@ -1,7 +1,7 @@
 public class Location{
-  public short x, y;
+  public int x, y;
 
-  public Location(short x_, short y_) {
+  public Location(int x_, int y_) {
     x = x_;
     y = y_;
   }
